# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Dialogue Time API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "https://api.dialogueone.com",
  "basePath": "/api/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/Schedule/start/{start}/end/{end}/tables": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "Retrieves all schedule records between two dates that have a table booking",
        "parameters": [
          {
            "name": "start",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}